move authOptions
