+++ src/app/[category]/[productName]/page.tsx
@@ -12,7 +12,7 @@ import Link from "next/link";
 import { notFound } from "next/navigation";
 import AddToCart from "./_components/AddToCart";
 import { getServerSession } from "next-auth";
-import { authOptions } from "@/app/api/auth/[...nextauth]/route";
+import { authOptions } from "@/app/api/auth/[...nextauth]/authOptions";
 
 type ProductPageProps = {
   params: { productName: string };
