+++ src/app/page.tsx
@@ -3,7 +3,7 @@ import Categories from "@/components/Home/Categories";
 import Hero from "@/components/Home/Hero";
 import Products from "@/components/Home/Products";
 import { getServerSession } from "next-auth";
-import { authOptions } from "./api/auth/[...nextauth]/route";
+import { authOptions } from "./api/auth/[...nextauth]/authOptions";
 
 export default async function HomePage() {
   return (
